Adding new calibration class, RunCalibrationHandler, with parameters that are constant per run, not indexed by detector or DAQ elements. Adding code to handle EPICS data shift with respect to DAQ events, loading from DB. Adding the constants for BDXminiPass1

+++ src/libraries/EPICS/epicsData.cc
@@ -4,7 +4,7 @@
 #include <sstream>      // std::istringstream
 
 /*raw Data is in the format: value description \n */
-void epicsData::decode(string rawData,int deltaTime) {
+void epicsData::decode(string rawData,long int deltaTime) {
 	std::istringstream stream;
 	stream.str(rawData);
 	double val;
